Use the make question option to avoid rebuild on macOS
There was already the earlier strange ` before the
"unity.a' is up to date message on macOS.

Turns out it might still not work on macOS because of certain
verbosity levels.

So use the question option instead of dry-run + grep as the
latter is more brittle, and the former is intended for this
purpose.